pyproject.toml
[tool.poetry]
name = "testebrmed"
version = "0.1.0"
description = ""
authors = ["Volney Casas <volneyrock@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
requests = "^2.31.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
